A working overlay for the ethernet module.
This overlay works with the BBB and the ethernet module, though still
working on getting the interface usable. According to syslog the
module's driver gets loaded successfully.

I need to check on the pinmux modes being used here. Everything works in
MODE3, but I thought I needed things in MODE7 to work.