Updating interrupt guess.
I think that the interrupt parent, gpio[0-3] corresponds to the gpio
class's that exists in /sys/class/gpio/. With this, I can then determine
what the interrupt should be in the devicetree.

E.g., P9.25 corresponds to gpio3[21] per
https://github.com/nomel/beaglebone/blob/master/gpio-header/generated/gpio-P9.25-00A0.dts,
so I think I can make the interrupt-parent gpio3 and the interrupt
number 21.